+++ mobile/lib/models/symptom_value_model.dart
@@ -2,7 +2,7 @@ class SymptomDetails {
   final int id;
   final String symptomName;
   final String symptomType;
-  final int symptomValue;
+  final double symptomValue;
 
   SymptomDetails({
     required this.id,
